Create README.md
I used to be a developer, started coding when I was 13 years old, went to technical high school, and graduated in computer science, but diverted when going to college. Took an engineering degree, did some coding in college, but that was a long time ago. Recently I decided I should resume my software engineering studies and here I am. Back to square 0.1